% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_parameters.R
\name{atu}
\alias{atu}
\title{Create ATU Target Parameter}
\usage{
atu(dgp, l = 1)
}
\arguments{
\item{dgp}{Data Generating Process object.}

\item{l}{Model index (default is 1).}
}
\value{
A TargetParameter object for ATT.
}
\description{
Create ATU Target Parameter
}
