% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_parameters.R
\name{compute_gamma_star}
\alias{compute_gamma_star}
\title{Compute Gamma Star for Each Model}
\usage{
compute_gamma_star(target.parameter, bases, dgp)
}
\arguments{
\item{target.parameter}{Target parameter object.}

\item{bases}{List of basis functions for each model.}

\item{dgp}{Data Generating Process object.}
}
\value{
A list of Gamma_star matrices for each model.
}
\description{
Compute Gamma Star for Each Model
}
