% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_parameters.R
\name{compute_population_bounds}
\alias{compute_population_bounds}
\title{Compute Target Parameter Population Bounds}
\usage{
compute_population_bounds(
  target.parameter = c("AUO", "ATO", "ATE", "ATT", "ATU", "LATE"),
  dgp,
  bases,
  late.lb,
  late.ub,
  assumptions,
  assumptions.extra
)
}
\arguments{
\item{target.parameter}{A vector of target parameters.}

\item{dgp}{Data Generating Process object.}

\item{bases}{A list of basis functions.}

\item{late.lb}{Lower limit for LATE.}

\item{late.ub}{Upper limit for LATE.}

\item{assumptions}{A list of assumptions to be considered. Default is NULL.}

\item{assumptions.extra}{A list of supplementary parameters that go along with the assumptions. Default is NULL.}
}
\value{
The population value for each target parameter.
}
\description{
Compute Target Parameter Population Bounds
}
