% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_parameters.R
\name{compute_population_value}
\alias{compute_population_value}
\title{Compute Target Parameter Population Values}
\usage{
compute_population_value(
  target.parameter = c("AUO", "ATO", "ATE", "ATT", "ATU", "LATE"),
  dgp,
  late.lb = NULL,
  late.ub = NULL
)
}
\arguments{
\item{target.parameter}{A vector of target parameters.}

\item{dgp}{Data Generating Process object.}

\item{late.lb}{Lower limit for LATE.}

\item{late.ub}{Upper limit for LATE.}
}
\value{
The population value for each target parameter.
}
\description{
Compute Target Parameter Population Values
}
