% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp}
\alias{dgp}
\title{Create DGP Object for MST2018 Model}
\usage{
dgp(suppZ = NULL, densZ = NULL, pscoreZ = NULL, mtrs = NULL, MST2018 = FALSE)
}
\arguments{
\item{suppZ}{Support of Z.}

\item{densZ}{Density of Z.}

\item{pscoreZ}{Propensity score of Z.}

\item{mtrs}{List of Marginal Treatment Response (MTR) functions.}

\item{MST2018}{A Boolean indicating whether to use the DGP in Mogstad, Santos, and Torgovitsy (2018).}
}
\value{
A new DGP object.
}
\description{
This function creates a DGP object based on the MST2018 model.
}
\examples{
basis0 = bernstein_basis(2)
theta0 = c(0.6, 0.4, 0.3)
mtr0 = MTR(basis0, theta0)

basis1 = bernstein_basis(2)
theta1 = c(0.75, 0.5, 0.25)
mtr1 = MTR(basis1, theta1)

dgp(suppZ = 0:2,
    densZ = c(0.5, 0.4, 0.1),
    pscoreZ = c(0.35, 0.6, 0.7),
    mtrs = list(mtr0, mtr1))

}
