% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{expand_weights_df}
\alias{expand_weights_df}
\title{Expand Weights from Weight Data Frame}
\usage{
expand_weights_df(u, weights.df)
}
\arguments{
\item{u}{Treatment indicator (0 or 1).}

\item{weights.df}{Data frame containing weights.}
}
\value{
A data frame containing u values and their respective average weights when
d = 0, 1.
}
\description{
The \code{compute_average_weights()} and \code{compute_average_weights_ivlike()} functions
return dataframes with four columns: \code{uStart}, \code{uEnd}, \code{avgWeightD0}, and \code{avgWeightD1}.
This function accepts a dataframe (weights.df) and vector (u), and returns the average
weight when d = 0, 1 for specific u values.
}
