% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_weights}
\alias{get_weights}
\title{Get Weights from Weight Data Frame}
\usage{
get_weights(weights.df, D = c(0, 1), nonZero = TRUE)
}
\arguments{
\item{weights.df}{Data frame containing weights.}

\item{D}{Treatment indicator (0 or 1).}

\item{nonZero}{Accepts logical indicating whether to output only non-zero weights
or all weights.}
}
\value{
A data frame containing weights.
}
\description{
Get Weights from Weight Data Frame
}
