% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivprte-package.R
\docType{package}
\name{ivprte-package}
\alias{ivprte}
\alias{ivprte-package}
\title{ivprte: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ian-xu-economics/ivprte}
  \item Report bugs at \url{https://github.com/ian-xu-economics/ivprte/issues}
}

}
\author{
\strong{Maintainer}: Ian Xu \email{ianxu@uchicago.edu} [copyright holder]

}
\keyword{internal}
