% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{compute_average_weights}
\alias{compute_average_weights}
\title{Compute Average Weights for Target Parameters}
\usage{
compute_average_weights(
  tp = c("AUO", "ATO", "ATE", "ATT", "ATU", "LATE"),
  dgp = NULL,
  data = NULL,
  late.lb = NULL,
  late.ub = NULL
)
}
\arguments{
\item{tp}{The target parameter to compute the average weights for.}

\item{dgp}{An optional DGP object containing the data generating process.}

\item{data}{An optional data.frame from which the estimated weights are calculated from.}

\item{late.lb}{Required when \code{tp} is set to "LATE", specifies the lower bound for integration.}

\item{late.ub}{Required when \code{tp} is set to "LATE", specifies the upper bound for integration.}
}
\value{
A data frame containing the average weights.
}
\description{
This function computes the average weights for different target parameters.
}
